refactor : 폴더 위치 변경 및 폴더 이름 수정 리팩토링
