+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductFacade.java
@@ -4,8 +4,8 @@ package com.loopers.application.product;
 import com.loopers.application.common.PageInfo;
 import com.loopers.application.product.dto.ProductCriteria;
 import com.loopers.application.product.dto.ProductInfo;
-import com.loopers.cache.ProductDetailCache;
-import com.loopers.cache.ProductListCache;
+import com.loopers.cache.product.ProductDetailCache;
+import com.loopers.cache.product.ProductListCache;
 import com.loopers.confg.kafka.KafkaMessage;
 import com.loopers.domain.brand.BrandModel;
 import com.loopers.domain.brand.BrandService;
