+++ apps/commerce-api/src/main/java/com/loopers/cache/product/ProductDetailCache.java
@@ -1,4 +1,4 @@
-package com.loopers.cache;
+package com.loopers.cache.product;
 
 
 import com.loopers.application.product.dto.ProductInfo;
