+++ apps/commerce-api/src/main/java/com/loopers/cache/product/ProductListKeyBuilder.java
@@ -1,4 +1,4 @@
-package com.loopers.cache;
+package com.loopers.cache.product;
 
 import org.springframework.stereotype.Component;
 import com.loopers.interfaces.api.product.OrderType;
