+++ apps/commerce-api/src/test/java/com/loopers/application/product/ProductListCacheTest.java
@@ -4,10 +4,10 @@ package com.loopers.application.product;
 import com.loopers.application.common.PageInfo;
 import com.loopers.application.product.dto.ProductCriteria;
 import com.loopers.application.product.dto.ProductInfo;
-import com.loopers.cache.ProductCachePolicyRegistry;
-import com.loopers.cache.ProductLikeVersionService;
-import com.loopers.cache.ProductListCache;
-import com.loopers.cache.ProductListKeyBuilder;
+import com.loopers.cache.product.ProductCachePolicyRegistry;
+import com.loopers.cache.product.ProductLikeVersionService;
+import com.loopers.cache.product.ProductListCache;
+import com.loopers.cache.product.ProductListKeyBuilder;
 import com.loopers.interfaces.api.product.OrderType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
